Restructure adapters internally
